Increase data type sizes in struct protocol_t

+++ libs/pilight/libs/pilight/protocols/protocol.h
@@ -72,11 +72,11 @@ typedef struct protocol_devices_t {
 
 typedef struct protocol_t {
   char *id;
-  uint8_t rawlen;
-  uint8_t minrawlen;
-  uint8_t maxrawlen;
-  uint16_t mingaplen;
-  uint16_t maxgaplen;
+  uint16_t rawlen;
+  uint16_t minrawlen;
+  uint16_t maxrawlen;
+  uint32_t mingaplen;
+  uint32_t maxgaplen;
   uint8_t txrpt;
   uint8_t rxrpt;
   //short multipleId;
